feat(decode_error): added field_exists fn() to check recursively for sources
